Added error message when exception occurs in main event loop
* Qt doesn't want exceptions to occur during handling events in the
event loop. Python filters give callbacks that may throw to C++
code, and may be called during the event loop handling. To fix this
we reimplement QApplication::notify and catch the exceptions there
and then delegate to the base implementation. If we catch an
exception we exit the application because we cannot guarantee the
state of the application being valid.

Signed-off-by: Jared Duffey <jared.duffey@bluequartz.net>